requirements.txt
click==7.1.2
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.0.1
MarkupSafe==2.0.1
numpy==1.20.2
pandas==1.4.0
pickleshare==0.7.5
prompt-toolkit==3.0.24
pure-eval==0.2.2
Pygments==2.11.2
python-dateutil==2.8.1
pytz==2021.1
retrying==1.3.3
scikit-learn==0.24.1
scipy==1.6.2
setuptools==57.4.0
six==1.15.0
stack-data==0.1.4
threadpoolctl==2.1.0
tomli==1.2.3
traitlets==5.1.1
typing_extensions==4.0.1
wcwidth==0.2.5
Werkzeug==2.0.2
xgboost==1.5.2